Move polygon edit modal from RegionMap to GeoModal
